requirements.txt
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.18
django-bootstrap4==0.0.8
django-heroku==0.3.1
django-registration==2.4.1
djangorestframework==3.9.4
gunicorn==19.9.0
Pillow==6.0.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.1
whitenoise==4.1.2